Fixed: Added protection against the short time logs being erased due to a silent failure to get the associated preference value.

+++ domoticz/main/SQLHelper.cpp
@@ -3994,41 +3994,49 @@ void CSQLHelper::AddCalendarUpdateUV()
 void CSQLHelper::CleanupShortLog()
 {
 	int n5MinuteHistoryDays=1;
-	GetPreferencesVar("5MinuteHistoryDays", n5MinuteHistoryDays);
+	if(GetPreferencesVar("5MinuteHistoryDays", n5MinuteHistoryDays))
+    {
+        // If the history days is zero then all data in the short losgs is deleted!
+        if(n5MinuteHistoryDays == 0)
+        {
+            _log.Log(LOG_ERROR,"CleanupShortLog(): MinuteHistoryDays is zero!");
+            return;
+        }
 
-	time_t now = mytime(NULL);
-	now-=(n5MinuteHistoryDays*24*3600);
-	struct tm tm2;
-	localtime_r(&now,&tm2);
+	    time_t now = mytime(NULL);
+	    now-=(n5MinuteHistoryDays*24*3600);
+	    struct tm tm2;
+	    localtime_r(&now,&tm2);
 
-	if (tm2.tm_year==70)
-	{
-		//Lost our time?
-		return;
-	}
+	    if (tm2.tm_year==70)
+	    {
+		    //Lost our time?
+		    return;
+	    }
 
-	char szDateStr[40];
-	char szTmp[200];
+	    char szDateStr[40];
+	    char szTmp[200];
 
-	sprintf(szDateStr,"%04d-%02d-%02d %02d:%02d:00",tm2.tm_year+1900,tm2.tm_mon+1,tm2.tm_mday,tm2.tm_hour,tm2.tm_min);
+	    sprintf(szDateStr,"%04d-%02d-%02d %02d:%02d:00",tm2.tm_year+1900,tm2.tm_mon+1,tm2.tm_mday,tm2.tm_hour,tm2.tm_min);
 
-	sprintf(szTmp,"DELETE FROM Temperature WHERE (Date<'%s')",szDateStr);
-	query(szTmp);
+	    sprintf(szTmp,"DELETE FROM Temperature WHERE (Date<'%s')",szDateStr);
+	    query(szTmp);
 
-	sprintf(szTmp,"DELETE FROM Rain WHERE (Date<'%s')",szDateStr);
-	query(szTmp);
+	    sprintf(szTmp,"DELETE FROM Rain WHERE (Date<'%s')",szDateStr);
+	    query(szTmp);
 
-	sprintf(szTmp,"DELETE FROM Wind WHERE (Date<'%s')",szDateStr);
-	query(szTmp);
+	    sprintf(szTmp,"DELETE FROM Wind WHERE (Date<'%s')",szDateStr);
+	    query(szTmp);
 
-	sprintf(szTmp,"DELETE FROM UV WHERE (Date<'%s')",szDateStr);
-	query(szTmp);
+	    sprintf(szTmp,"DELETE FROM UV WHERE (Date<'%s')",szDateStr);
+	    query(szTmp);
 
-	sprintf(szTmp,"DELETE FROM Meter WHERE (Date<'%s')",szDateStr);
-	query(szTmp);
+	    sprintf(szTmp,"DELETE FROM Meter WHERE (Date<'%s')",szDateStr);
+	    query(szTmp);
 
-	sprintf(szTmp,"DELETE FROM MultiMeter WHERE (Date<'%s')",szDateStr);
-	query(szTmp);
+	    sprintf(szTmp,"DELETE FROM MultiMeter WHERE (Date<'%s')",szDateStr);
+	    query(szTmp);
+    }
 }
 
 void CSQLHelper::DeleteHardware(const std::string &idx)
